added sockclient.cpp
